easy coding -
recap & re-organized about 对象是否相等；

map中get(key)时，如何判断key是否存在在散列表中；
自定义类型的判等 - equals() + hashCode();
list是否相同 - 取决于equals()方法的具体实现。

Objects类的静态方法会很有用
